Streamline ParserGetVariableValueSwitch hook handler signature
Most of the & are not needed. They are only there for compatibility
with extremely old PHP versions, from a time when PHP was passing
objects not by ref but by cloning them. It doesn't make sense to have
code that requires the Parser as a reference, like it want's to replace
the parser with a different one.

I'm also removing the strict type hints from arguments that are not
used in this code. This does make this code more robuts. It can't start
failing when the type of an argument this code doesn't even care about
changes.

Change-Id: I6b8287ec626d165fcaa6888b44bef273101ff987